+++ core/data/src/main/kotlin/com/unifest/android/core/data/repository/FestivalRepository.kt
@@ -2,15 +2,10 @@ package com.unifest.android.core.data.repository
 
 import com.unifest.android.core.model.FestivalModel
 import com.unifest.android.core.model.FestivalTodayModel
-import kotlinx.coroutines.flow.Flow
 
 interface FestivalRepository {
     suspend fun getAllFestivals(): Result<List<FestivalModel>>
     suspend fun searchFestival(name: String): Result<List<FestivalModel>>
     suspend fun getIncomingFestivals(): Result<List<FestivalModel>>
     suspend fun getTodayFestivals(date: String): Result<List<FestivalTodayModel>>
-    fun getLikedFestivals(): Flow<List<FestivalModel>>
-    suspend fun insertLikedFestivalAtHome(festival: FestivalTodayModel)
-    suspend fun insertLikedFestivalAtSearch(festival: FestivalModel)
-    suspend fun deleteLikedFestival(festival: FestivalModel)
 }
